Update entrypoint.sh
Optionally run pip3 install when file is present.

+++ entrypoint.sh
@@ -2,6 +2,8 @@
 
 set -e
 
-pip3 install -r requirements.txt
+if [ ! -f "requirements.txt" ]; then
+  pip3 install -r requirements.txt
+fi
 
 sh -c "cdk $*"
